pyproject.toml
[project]
name = "crawjud-reestruturado"
version = "0.1.0"
description = ""
authors = [
    {name = "Robotz213",email = "nicholas@robotz.dev"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "quart (>=0.20.0,<0.21.0)",
    "quart-jwt-extended (>=0.1.0,<0.2.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "quart-cors (>=0.8.0,<0.9.0)",
    "selenium (>=4.32.0,<5.0.0)",
    "webdriver-manager (>=4.0.2,<5.0.0)",
    "pypdf (>=5.5.0,<6.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "pytz (>=2025.2,<2026.0)",
    "openai (>=1.78.1,<2.0.0)",
    "celery (>=5.5.2,<6.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "google-cloud-storage (>=3.1.0,<4.0.0)",
    "google-auth (>=2.40.1,<3.0.0)",
    "redis (>=3.5.3,<6.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "hypercorn (>=0.17.3,<0.18.0)",
    "clear (>=2.0.0,<3.0.0)",
    "python-socketio (>=5.13.0,<6.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "aiohttp (>=3.11.18,<4.0.0)",
    "cryptography (>=45.0.2,<46.0.0)",
    "quart-flask-patch (>=0.3.0,<0.4.0)",
    "quart-socketio @ git+https://github.com/Robotz213/Quart-SocketIO.git",
    "flask-session[redis] (>=0.8.0,<0.9.0)",
    "python-magic (>=0.4.27,<0.5.0)",
    "minio (>=7.2.15)",
    "rsa (>=4.7)",
    "inquirer (>=3.4.0,<4.0.0)",
    "redis-om (>=0.0.20)",
    "types-redis (>=4.1.17)",
    "jpype1 (>=1.6.0)",
    "pytesseract (>=0.3.13,<0.4.0)",
    "opencv-python (>=4.12.0.88,<5.0.0.0)",
    "debugpy (>=1.8.15,<2.0.0)",
    "python-docx (>=1.2.0,<2.0.0)",

]
[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"
quart-jwt-extended = "^0.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
